+++ src/js/app/intlProvider.js
@@ -4,21 +4,22 @@ import dot from 'dot-object'
 
 import ComponentInjector from './injector'
 
-import en from 'react-intl/locale-data/en'
-import de from 'react-intl/locale-data/de'
-
 import deMessages from './i18n/de.js'
 import enMessages from './i18n/en.js'
 
-addLocaleData([...en, ...de])
-
-const messages = {
-    en: enMessages,
-    de: deMessages,
-}
-
 export default ComponentInjector.return('App.IntlProvider', connect(
     (globalState) => {
+        const { addLocaleData } = require('react-intl')
+        const en = require('react-intl/locale-data/en')
+        const de = require('react-intl/locale-data/de')
+
+        addLocaleData([...en, ...de])
+
+        const messages = {
+            en: enMessages,
+            de: deMessages,
+        }
+
         let language = globalState.app.context.getLanguage()
 
         return {