+++ src/php/FrontendBundle/EventListener/ContentSecurityPolicy.php
@@ -5,11 +5,9 @@ namespace Frontastic\Catwalk\FrontendBundle\EventListener;
 use Frontastic\Common\ReplicatorBundle\Domain\Project;
 use Symfony\Component\HttpKernel\Event\FilterResponseEvent;
 
-use Frontastic\Catwalk\ApiCoreBundle\Domain\Context;
-
 class ContentSecurityPolicy
 {
-    const DEFAULT = [
+    const BASELINE = [
         'default-src' => [
             'self',
         ],
@@ -17,8 +15,6 @@ class ContentSecurityPolicy
             'self',
             'unsafe-inline',
             'unsafe-eval',
-            'https://ssl.google-analytics.com',
-            'https://connect.facebook.net',
         ],
         'img-src' => [
             'self',
@@ -35,8 +31,11 @@ class ContentSecurityPolicy
         ],
         'frame-src' => [
             'self',
-            'https://www.facebook.com',
-            'https://s-static.ak.facebook.com',
+        ],
+        'frame-ancestors' => [
+            'self',
+            'https://*.frontastic.io',
+            'frontastic.io.local',
         ],
         'object-src' => [
             'self',
@@ -45,6 +44,17 @@ class ContentSecurityPolicy
             'self',
             'ws:',
             'wss:',
+        ],
+    ];
+
+    const DEFAULT = [
+        'script-src' => [
+            'https://ssl.google-analytics.com',
+            'https://connect.facebook.net',
+        ],
+        'frame-src' => [
+            'https://www.facebook.com',
+            'https://s-static.ak.facebook.com',
         ],
     ];
 
@@ -61,7 +71,10 @@ class ContentSecurityPolicy
     public function onKernelResponse(FilterResponseEvent $event)
     {
         $responseHeaders = $event->getResponse()->headers;
-        $policies = $this->project->configuration['policy'] ?? self::DEFAULT;
+        $policies = $this->merge(
+            self::BASELINE,
+            $this->project->configuration['policy'] ?? self::DEFAULT
+        );
 
         $responseHeaders->set(
             'Content-Security-Policy',
@@ -82,7 +95,7 @@ class ContentSecurityPolicy
 
                                         return $target;
                                     },
-                                    $values
+                                    array_unique($values)
                                 )
                             )
                         );
@@ -92,5 +105,10 @@ class ContentSecurityPolicy
                 )
             )
         );
+    }
+
+    private function merge($policy, $additions): array
+    {
+        return array_merge_recursive($policy, $additions);
     }
 }